add easing to chart
