+++ charming/src/element/easing.rs
@@ -1,0 +1,38 @@
+use serde::Serialize;
+
+#[derive(Clone, Debug, Default, Serialize, Copy, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub enum Easing {
+    Linear,
+    QuadraticIn,
+    QuadraticOut,
+    QuadraticInOut,
+    CubicIn,
+    #[default]
+    CubicOut,
+    CubicInOut,
+    QuarticIn,
+    QuarticOut,
+    QuarticInOut,
+    QuinticIn,
+    QuinticOut,
+    QuinticInOut,
+    SinusoidalIn,
+    SinusoidalOut,
+    SinusoidalInOut,
+    ExponentialIn,
+    ExponentialOut,
+    ExponentialInOut,
+    CircularIn,
+    CircularOut,
+    CircularInOut,
+    ElasticIn,
+    ElasticOut,
+    ElasticInOut,
+    BackIn,
+    BackOut,
+    BackInOut,
+    BounceIn,
+    BounceOut,
+    BounceInOut,
+}
