+++ charming/src/element/mod.rs
@@ -17,6 +17,7 @@ pub mod coordinate;
 pub mod coordinate_tooltip;
 pub mod cursor;
 pub mod dimension_encode;
+pub mod easing;
 pub mod emphasis;
 pub mod font_settings;
 pub mod formatter;
@@ -69,6 +70,7 @@ pub use coordinate::*;
 pub use coordinate_tooltip::*;
 pub use cursor::*;
 pub use dimension_encode::*;
+pub use easing::*;
 pub use emphasis::*;
 pub use formatter::*;
 pub use icon::*;
